BuildersModule along with ContributesAndroidInjector to be in sync with modern Dagger 2 practices. Converting MainActivity to use DI and other goodies.
