+++ app/build.gradle
@@ -43,6 +43,10 @@ android {
     }
 }
 
+androidExtensions {
+    experimental = true
+}
+
 repositories {
     maven {
         url "https://maven.google.com"
@@ -57,13 +61,13 @@ dependencies {
     implementation 'com.android.support:gridlayout-v7:27.1.0'
     implementation 'com.android.support:design:27.1.0'
     implementation "org.jetbrains.anko:anko:$anko_version"
-    implementation 'com.google.firebase:firebase-database:11.8.0'
-    implementation 'com.google.firebase:firebase-core:11.8.0'
-    implementation 'com.google.firebase:firebase-firestore:11.8.0'
+    implementation 'com.google.firebase:firebase-database:12.0.0'
+    implementation "com.google.firebase:firebase-core:$firebase_version"
+    implementation "com.google.firebase:firebase-firestore:$firebase_version"
     implementation 'com.firebaseui:firebase-ui-firestore:3.2.2'
-    implementation 'com.google.firebase:firebase-perf:11.8.0'
+    implementation "com.google.firebase:firebase-perf:$firebase_version"
     implementation 'com.jaredrummler:material-spinner:1.1.0'
-    implementation 'android.arch.lifecycle:extensions:1.1.0'
+    implementation 'android.arch.lifecycle:extensions:1.1.1'
     implementation 'com.squareup.picasso:picasso:2.5.2'
     implementation ('com.crashlytics.sdk.android:crashlytics:2.8.0@aar') {
         transitive = true
@@ -72,8 +76,10 @@ dependencies {
     implementation 'com.google.dagger:dagger:2.15'
     implementation 'com.google.dagger:dagger-android:2.15'
     implementation 'com.google.dagger:dagger-android-support:2.15'
+    implementation 'com.jakewharton:butterknife:8.8.1'
     kapt 'com.google.dagger:dagger-compiler:2.15'
     kapt 'com.google.dagger:dagger-android-processor:2.15'
+    kapt 'com.jakewharton:butterknife-compiler:8.8.1'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support:support-annotations:27.1.0'
     androidTestImplementation 'com.android.support.test:rules:1.0.1'
@@ -83,6 +89,10 @@ dependencies {
 
     implementation 'com.android.support.constraint:constraint-layout:1.0.2'
     implementation 'com.android.support:cardview-v7:27.1.0'
+}
+
+kapt {
+    generateStubs = true
 }
 
 apply plugin: 'com.google.gms.google-services'
