+++ app/src/main/java/re/spitfy/ctftime/activities/BaseActivity.kt
@@ -1,0 +1,5 @@
+package re.spitfy.ctftime.activities
+
+import android.support.v7.app.AppCompatActivity
+
+abstract class BaseActivity : AppCompatActivity()
