--- app/src/main/java/re/spitfy/ctftime/di/FirestoreModule.kt
@@ -1,14 +1,0 @@
-package re.spitfy.ctftime.di
-
-import com.google.firebase.firestore.FirebaseFirestore
-import dagger.Module
-import dagger.Provides
-import javax.inject.Singleton
-
-@Module
-class FirestoreModule {
-
-    @Provides
-    @Singleton
-    fun provideFirestore(): FirebaseFirestore = FirebaseFirestore.getInstance()
-}
