+++ app/src/main/java/re/spitfy/ctftime/di/MainActivityModule.kt
@@ -1,21 +1,6 @@
 package re.spitfy.ctftime.di
 
-import android.app.Activity
-import dagger.Binds
 import dagger.Module
-import dagger.android.ActivityKey
-import dagger.android.AndroidInjector
-import dagger.multibindings.IntoMap
-import re.spitfy.ctftime.activities.MainActivity
 
-@Module(subcomponents = [MainActivitySubcomponent::class])
-abstract class MainActivityModule {
-
-    @Binds
-    @IntoMap
-    @ActivityKey(MainActivity::class)
-    abstract fun bindMainActivityInjectorFactory(
-        builder: MainActivitySubcomponent.Builder
-    ): AndroidInjector.Factory<out Activity>
-
-}
+@Module(subcomponents = [MainActivitySubComponent::class])
+abstract class MainActivityModule
