+++ app/src/main/java/re/spitfy/ctftime/di/app/ApplicationComponent.kt
@@ -1,0 +1,30 @@
+package re.spitfy.ctftime.di.app
+
+import android.app.Application
+import android.content.Context
+import com.google.firebase.firestore.CollectionReference
+import com.google.firebase.firestore.FirebaseFirestore
+import dagger.BindsInstance
+import dagger.Component
+import dagger.android.support.AndroidSupportInjectionModule
+import re.spitfy.ctftime.CTFTimeApp
+import re.spitfy.ctftime.di.MainActivityModule
+import javax.inject.Singleton
+
+@Singleton
+@Component(modules = [
+    AndroidSupportInjectionModule::class,
+    AppModule::class,
+    BuildersModule::class]
+)
+interface ApplicationComponent {
+
+    @Component.Builder
+    interface Builder {
+        @BindsInstance fun application(application: Application): Builder
+        fun build(): ApplicationComponent
+    }
+
+    fun inject(app: CTFTimeApp)
+
+}
