+++ app/src/main/java/re/spitfy/ctftime/di/app/BuildersModule.kt
@@ -1,0 +1,14 @@
+package re.spitfy.ctftime.di.app
+
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+import re.spitfy.ctftime.activities.MainActivity
+import re.spitfy.ctftime.di.MainActivityModule
+
+@Module
+abstract class BuildersModule {
+
+    @ContributesAndroidInjector(modules = [MainActivityModule::class])
+    abstract fun bindMainActivity(): MainActivity
+
+}
