+++ build.gradle
@@ -3,12 +3,14 @@
 buildscript {
     ext.kotlin_version = '1.2.31'
     ext.anko_version = '0.10.4'
+    ext.firebase_version = '12.0.0'
 
     repositories {
         google()
         jcenter()
         maven { url 'https://maven.fabric.io/public' }
     }
+
     dependencies {
         classpath 'com.android.tools.build:gradle:3.0.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
